# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    name: IT Support
  description: IT Broadcast Operations & Helpdesk System API. Provides endpoints for
    ticket management, knowledge base (Big Book), shift scheduling, and real-time
    notifications.
  license:
    name: MIT
  title: G-Report API
  version: "1.0"
paths:
  /auth/login:
    get:
      description: Display the login page
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
      summary: Show login form
      tags:
      - Auth
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authenticate user with email/username and password
      parameters:
      - description: Email or Username
        in: formData
        name: email
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - text/html
      responses:
        "302":
          description: Redirect to dashboard
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
      summary: User login
      tags:
      - Auth
...
  /consumer:
    get:
      description: Display consumer dashboard with recent tickets
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Consumer dashboard
      tags:
      - Consumer
...
  /manager:
    get:
      description: Display manager dashboard with KPIs, charts, and management tools
      parameters:
      - description: Month filter (YYYY-MM)
        in: query
        name: month
        type: string
      - description: Pagination page for resolved tickets
        in: query
        name: history_page
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Manager dashboard
      tags:
      - Manager
...
  /staff:
    get:
      description: Display staff dashboard with active tickets and routines
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Staff dashboard
      tags:
      - Staff
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: user_id
    type: apiKey
swagger: "2.0"